DAS-2180: Extracts harmony_service_entry.
First stab at separating concerns. This now has a
harmony_browse_image_generator as well as a harmony_service_entry.
