+++ docker/service_version.txt
@@ -1,1 +1,1 @@
-1.2.2
+1.3.0
